rename selenium folder to "pages"
